+++ forge/src/main/java/net/mehvahdjukaar/polytone/forge/PolytoneForge.java
@@ -20,7 +20,7 @@ public class PolytoneForge {
 
     public PolytoneForge() {
         if (FMLEnvironment.dist == Dist.CLIENT) {
-            Polytone.init(false);
+            Polytone.init(false, !FMLEnvironment.production);
 
             MinecraftForge.EVENT_BUS.register(this);
         } else {